Implemented JSON thread parser
Successfully parsed the JSON containing all the subscriptions for the user and stored each category into an arraylist. Also modified the session manager so that it returns the password if given the key. This way all the parsing can be done in one activity (Main Activity) so it'll avoid passing data in between activites. As a result, it'll avoid relaunch crashes while session is active, since data could be null on relaunch.