Implemented custom adapter for displaying user info in profile activity
Added a recycler view containing all the user info, like email, tribute points, subscriptions, etc. Also implemented a custom adapter for it in the user profile, so that different imageviews/icons can be inflated for each item. Also allows the user in the future to scroll through the view, as more information gets added.