Implemented system for updating announcements
Connected announcements with the backend and admins can now create the messages through the server so that all users will receive the updates. That way we can control what we want to say to the community without having to make them download a newer version of the app. Created the announcement hander, in which it'll echo back the message we set in the backend. As of now, it takes no parameters for the data posted.