Implemented caching mechanism using Picasso + okhttp
Thread images are now being cached on the device using Picasso and okhttp. Debug indicators are enabled right now and the corner triangles show if the image is being loaded from network (red), disk (blue), or memory (green). Can disable indicators in the App.java. Can now use this same pattern for user images, and will be useful later if we add the user images to the chat activity.