Improved scrolling performance for viewpager
Removed calls on blocking methods that download thread images from the server. We no longer need to block the UI thread everytime to get the images since we can now retrieve the images from the cache, or through the server using okhttp and Picasso. No need for the handler and php now, but keep them just in case. The block on the UI thread everytime the fragment is refreshed is most likely what caused the lags when swiping. Maybe we can further optimize the performance later since there's still lag, but it's a lot better than before.