Uncached avatars + modified image sizes
Removed caching for user profile images since there's no way for the cache to update if a user changes their avatar, unless we make major architecture changes on the front and back end. The performance is still decent enough even if we have to download the avatars every time. Also modified the dimensions of the avatar images and removed all 'fit' attributes. Need to see if that fixes the rendering issue on different screen sizes. Also added avatar link to the session manager.