Replaced viewpager with tab layout buttons
Removed viewpager and used tab buttons instead. App can now filter onNew and onTrend when the user clicks on the tabs. Home tab shows all the threads by date, regardless of categories. But the user cant swipe since there's no viewpager, they can only click the buttons on the bottom. Now, we need to create new activity that slides in when user searches, or filters by category. However, right now the search is broken and doesn't search, and user can't scroll on a filtered fragment (the app completely freezes when it hits a loader).