Fix crash for logged in users + refactored code
Added in wifi checks in main activity so that the app won't crash if user launches the app while logged in, without being connected to the internet. The error dialog will force the user out of the app, so they can connect and then try again. Also refactored code to make it more organized and maintainable.