Stored recent user activity in DB + introduced Retrofit
App now tracks the user's recent comments in each room and stores it in the database. Later, we can use the data and create the UI for displaying the user's recent activity. Also used Retrofit library to make asynchronous network calls. This is a lot better and cleaner than the old way we used handlers, so we can potentially rewrite them using Retrofit when we have time.