Implemented permanent ban
Added new column to client users table, in which a 1 indicates that a certain user is banned. If a banned user tries to login the app, the server will tell the client side to display a message and stop the user from logging in. If we want to ban a user, we can just manually change the banned column in the database table using an update query.