Changed design for fetching categories + bug fix
Removed all parsing for categories/subscriptions. Also removed the field on the user profile since every user has the same categories anyways. Plus there was a visual bug on the user profile due to the parsing logic. The categories will now be read from the strings.xml instead of from the server. This is better design and also affects performance. Also sorted the categories in alphabetical order for better UX and navigation. In addition, I added a few more categories. Can add more later on. We can now remove the subscriptions field from the database and all the logic on the server side since we don't even need it anymore.