Implemented new search bar
When user clicks on magnifying glass icon on the top right corner of the screen, a search bar will appear below the tabs. User can hide the search bar by clicking on magnifying glass again, or by clicking the back button on the search bar (or even on your phone). By tapping/clicking the back button, it will hide the keyboard first. By tapping/clicking a second time, it will hide the search bar. This is much better UX compared to the first search bar we had for a couple reasons. The old search bar would take up the entire toolbar if clicked, thus hiding important info like the current filter. Also it had the same color as the toolbar which was hard to see. It was also hard for the user to exit that search bar since they had to click the x at the very top, which wasn't very visible either.