Implemented new spinner widget
Implemented new widget to replace spinner in post thread activity. User can now select category by clicking on a button, and then selecting an item from the dialog fragment. It's a lot easier to to navigate and click on. Will also revamp the post thread user flow and layout once the new wireframes are finished.