Implemented new user flow for post thread activity
Revamped the user flow for posting thread. Removed the toolbar post button, and created two pages for the post activity. First page has title, description, and category. Second and final page has image upload. Submit button is on the final page, and displays a confirmation dialog when user clicks on it. Also included a header UI section, with icons that indicate the user's progress in the user flow. The idea is to make the process easier and more intuitive for the user. Also encourages them to upload an image with the thread post, which is highly recommended.