Implemented FAQs section feature
FAQs section now available for user. Data is generated from server side so we can add/edit content by using the MySQL database. The recyclerview UI is expandable and collapsable upon tap. The user can go to the FAQs page by clicking on the entry point in the nav drawer.