Implemented EULA
Added End User License Agreement(EULA), due to the original app being taken down from the Google Playstore for not having a EULA and handling user sensitive data like email. Privacy section of the EULA should be able to cover for user data usage. User needs to accept EULA when first launching the app, in order to be able to use the app.